use locks in more places, give more control over thinks
esp discovered resources and lock file.
lock file can remain in place if process is interrupted, so a facility for deleting it is useful.
destroy and stop will automatically try to delete them.

the lock file seems to leak when AMP is terminated, probably because active shell processes are also terminated.

(note, this seems to be the main problem with locks being in use, NOT some of the places where we added the retryWithLock,
though SOME of those places are needed, so easier to keep them all.)

also make sure a failure to stop/destroy _will_ block shutdown